Replaced Markdown crate with Comrak
The Markdown crate says it has fully implemented CommonMark, but there
are issues, such as HRs not working properly. It also misses a bunch of
the extras commonly used in Markdown.

Pulldown-cmark is another crate worthy of consideration, and very widely
used, but it just doesn't seem to have the breadth of functionality
found in Comrak. Therefore Comrak seems worth using.

In usage there is practically no difference, beyond configuring Comrak
to enable all the extras required.