Added callouts to Markdown by extending blockquotes
Callouts are not a standard Markdown feature, and there are various
methods proposed to implement them, including single-cell tables, custom
markup, raw HTML, and blockquotes. Utilising blockquotes seems like the
best approach, as it is the closest in semantic association and also
degrades gracefully for anything that doesn't support the extended
functionality. They are still totally readable on GitHub for example.

The only problem is that there is no way to achieve the desired outcome
using pure CSS - so a small amount of JavaScript is necessitated in
order to locate the extended blockquotes and add an extra class to them.
Blockquotes are taken to be callouts if they start with a <strong>
element containing a single, specific word, such as "Warning". CSS can
then be used to target whichever of these is desired.

  - Added JavaScript to locate blockquotes that start with bold text,
    and contain a single word (i.e. no spaces). A classname is then
    added to the blockquote, which is the lowercase version of the word,
    with any punctuation or other non-word characters removed.

  - Added the JavaScript to the main layout template, but no others (as
    it's only needed when rendering Markdown pages).

  - Added routing to load the JavaScript file.

  - Added CSS styling for a range of callout types.