Removed use of hardlinks
Using hardlinks is ideal in order to save space, but an odd problem has
been encountered whereby at intermittent intervals the source files are
truncated. It is not readily reproducible, and also not immediately
apparent why this is happening. Therefore, for safety, use of hardlinks
has been removed, at least for now.