Added basic Dockerfile
Added a basic Dockerfile to help with building the application in a
container. This builds the application in two stages - first it creates
a build environment in which to prepare everything, and then the final
binary is transferred to a clean, minimal environment which is suitable
for distribution.

When evaluating options, debian:bullseye-slim resulted in a 94MB image
size (85MB once the executable was run through UPX), so distroless has
been chosen instead, as it gives a 36MB image size (27MB once run
through UPX).

The specific version of Rust has been set, rather than "latest", due to
Rust depending heavily on the version of glibc required. If there is a
change in a future version of Rust, and that version of glibc is not
available in the destination image, then it will not run. Setting the
version of Rust in use therefore prevents any surprise breakages.

It's worth noting that the host IP to serve on needs to be set to
0.0.0.0 to allow outside traffic to connect. In other words, the host
entry in Config.toml should be set to "0.0.0.0" for a Docker setup.