Added mold as an alternate linker
Mold is a much faster linker when building normally, i.e. outside of
Docker. Adding it to the build process has, oddly, not resulted in such
noticeable gains, with improvement being essentially unnoticeable, and
no more thand 5%. This merits further attention, but as it is generally
better to use mold than not, it has been kept in, for now at least.

  - Installed the mold linker and its dependencies.

  - Set cargo to use mold as the linker when building.