Changed distribution image to use Alpine
The distroless image is small, but by changing the build target and
compiling for musl, Alpine can be used. This results in a 21MB image
being produced, which becomes 12MB after running the executable through
UPX. There do not currently appear to be any side-effects from using
Alpine and musl.

The change to Alpine and musl means that a base of rust:latest can be
set, as using static compilation means there is no concern with future
versions of Rust not being compatible with the versions of glibc
available in the destination image.