Added configuration arguments
Some aspects of optimisation are useful to control by passing arguments
to docker build. The default position is to optimise for the smallest
possible image size, but this can result in longer compilation times.
Although this Dockerfile is intended for production builds, in some
cases speed may be favoured. Most of the options affecting this are
configured through the Cargo.toml file, but notably the use of UPX does
have a significant effect on time, and therefore an option has been
added to specifically control it. Additionally, it can be useful to
adjust build settings through command arguments and therefore another
argument has been added to pass options to cargo build.

For example:

docker build -t rustmark --build-arg upx=0 .

  - Added upx and cargo-opts arguments to the Dockerfile. Note that
    Docker determines the effect of argument declarations on layers
    based on where they are declared, not where they are used.

  - Added cargo-opts into the cargo build command.

  - Implemented a check around running UPX on the compiled binary. By
    default, this is enabled.

  - Separated out UPX to its own layer, as it is worth having an extra
    layer to gain the advantage of easily toggling it on and off
    without rebuilding the whole application.