Merged Docker layers
The attempt to achieve optimised staged builds has not worked out, as
the linking time has not been decreased, meaning that having a separate
content stage is of no benefit. The intention was that this would be a
quicker stage, but in reality there is no significant difference seen
when compared to the combined time - and having two stages means adding
the time of the additional stage, increasing the overall time with no
benefit.

  - Combined content stage into main codebase compilation stage.